Move encoder functions into extension of AWSShape
This allows us to tidy up the payload encoding
